requirements.txt
beautifultable==0.7.0
Click==7.0
Flask==1.0.3
graphviz==0.11
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.3.4
MarkupSafe==1.1.1
numpy==1.16.4
pandas==0.24.2
ParamUnittest==0.2
peewee==3.9.6
Pillow==6.0.0
python-dateutil==2.8.0
pytz==2019.1
selenium==3.141.0
six==1.12.0
torch==1.1.0
torchvision==0.3.0
torchviz==0.0.1
urllib3==1.25.3
Werkzeug==0.15.4
